refactor: moved GraphQL module "npms" resolvers to own type-based files
